Split Operator into its own module
